add more styling to sidebar
