+++ src/layouts/index.ts
@@ -1,3 +1,3 @@
 export { DashboardLayout } from './DashboardLayout';
-export { AuthLayout } from './AuthLayout';
+export * from './AuthLayout';
 export { BoardLayout } from './BoardLayout';
