+++ src/pages/Board/Board.styles.ts
@@ -1,0 +1,1 @@
+export const dummy = {};
