+++ src/pages/Board/Board.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { BoardLayout } from '../layouts';
+import { BoardLayout } from '../../layouts';
 
 const BoardPage: React.FC = () => {
   return <BoardLayout></BoardLayout>;
