+++ src/pages/Board/index.ts
@@ -1,0 +1,1 @@
+export { default as BoardPage } from './Board';
