+++ src/pages/Signup/index.ts
@@ -1,0 +1,1 @@
+export { default as SignupPage } from './Signup';
