+++ src/pages/index.ts
@@ -1,4 +1,4 @@
-export { default as LoginPage } from './Login';
-export { default as SignupPage } from './Signup';
-export { default as DashboardPage } from './Dashboard';
-export { default as BoardPage } from './Board';
+export { LoginPage } from './Login';
+export { SignupPage } from './Signup';
+export { DashboardPage } from './Dashboard';
+export { BoardPage } from './Board';
