+++ src/theme/themes.ts
@@ -7,9 +7,10 @@ export const defaultTheme = {
     header1: '1.8rem',
     header2: '1.6rem',
     header3: '1.4rem',
-    header4: '1.2rem',
+    header4: '1.15rem',
     header5: '1.1rem',
     paragraph: '1rem',
+    subtitle: '0.85rem',
     helperText: '0.8rem',
     copyrightText: '0.7rem',
   },
@@ -19,13 +20,15 @@ export const defaultTheme = {
       200: '#f4f5f7',
       300: '#e1e1e1',
       400: '#737581',
-      500: '#4a4b53',
-      600: '#000000',
+      500: '#6b788d',
+      600: '#4a4b53',
+      700: '#000000',
     },
     blue: {
       100: '#0079bf',
       200: '#026aa7',
-      300: '#172b4d',
+      300: '#42526e',
+      400: '#172b4d',
     },
     red: {
       100: '#ff4500',
